feat: SolutionInput component and business logic
Added a core folder holding business logic reusable methods.
Added a validateSolutions method that validates a solution and returns relevant data about validation results.

Added the SolutionInput component which takes advantage of the above method to validate a solution from the UI.

The SolutionInput component should be used with v-model to sync data.
Such widget automatically updates the model in the following way:

- If a valid solution is input, the model is updated with the provided solution.
- If an invalid solution is provided, the model is updated with NULL.

The solution business logic:

- Removes any newline sequence.
- Trims the input.
- Splits by empty space.
- Validate each item against the WCA moveset.
- Returns eloquent data about why the validation eventually failed.

The UI is not entirely completed, although the business logic
already provides enough information to build the UI.